extract Pool interface from UnboundPool
